CHANGE: moved auth helper files to separate folder
